refactor(checker): route namespace mismatch lookups
