+++ crates/tsz-checker/src/assignability/mod.rs
@@ -9,6 +9,7 @@ mod application_keyof_helpers;
 pub mod assignability_checker;
 mod assignability_diagnostics;
 mod assignability_eval;
+mod assignability_relation;
 mod assignability_type_param_helpers;
 pub mod assignment_checker;
 mod awaited_variance_normalization;
